refactor: share home assistant entity context formatting
