+++ internal/integrations/homeassistant/entitycontext/entity_format_vacuum.go
@@ -1,4 +1,4 @@
-package awareness
+package entitycontext
 
 import (
 	"time"
